refactor: 테스트 하드코딩 데이터 수정 및 패키지 변경
